+++ plugin-hrm-form/src/___tests__/utils/mappers.test.js
@@ -1,0 +1,44 @@
+import { mapCallType, mapChannel } from '../../utils';
+
+test('Test contact call type mapper', () => {
+  const mapSelf = 'Child calling about self';
+  const mapCaller = 'Someone calling about a child';
+  const string = 'anything else';
+
+  expect(mapCallType(mapSelf)).toEqual('SELF');
+  expect(mapCallType(mapCaller)).toEqual('CALLER');
+  expect(mapCallType(string)).toEqual('ANYTHING ELSE');
+});
+
+test('Test contact channel mapper', () => {
+  const ch1 = 'facebook';
+  const expectCh1 = 'Facebook Messenger';
+  const fmtCh1 = mapChannel(ch1);
+
+  const ch2 = 'web';
+  const expectCh2 = 'Chat';
+  const fmtCh2 = mapChannel(ch2);
+
+  const ch3 = 'voice';
+  const expectCh3 = 'Voice';
+  const fmtCh3 = mapChannel(ch3);
+
+  const ch4 = 'sms';
+  const expectCh4 = 'SMS';
+  const fmtCh4 = mapChannel(ch4);
+
+  const ch5 = 'whatsapp';
+  const expectCh5 = 'WhatsApp';
+  const fmtCh5 = mapChannel(ch5);
+
+  const undef = 'anything else';
+  const expectUndef = 'Undefined';
+  const fmtUndef = mapChannel(undef);
+
+  expect(fmtCh1).toEqual(expectCh1);
+  expect(fmtCh2).toEqual(expectCh2);
+  expect(fmtCh3).toEqual(expectCh3);
+  expect(fmtCh4).toEqual(expectCh4);
+  expect(fmtCh5).toEqual(expectCh5);
+  expect(fmtUndef).toEqual(expectUndef);
+});
